int8_t for error returning

+++ include/sn_syntax/sn_expr.h
@@ -11,28 +11,28 @@
 #define PrimType_Subscript      0x06
 #define PrimType_Call           0x07
 
-int scopes_expr(struct parser_st *, struct node_st *);
+int8_t scopes_expr(struct parser_st *, struct node_st *);
 
-int list_expr(struct parser_st *, struct node_st *);
+int8_t list_expr(struct parser_st *, struct node_st *);
 
-int ident_get_expr(struct parser_st *, struct node_st *);
+int8_t ident_get_expr(struct parser_st *, struct node_st *);
 
-int ident_new_expr(struct parser_st *, struct node_st *);
+int8_t ident_new_expr(struct parser_st *, struct node_st *);
 
-int bool_expr(struct parser_st *, struct node_st *);
+int8_t bool_expr(struct parser_st *, struct node_st *);
 
-int number_expr(struct parser_st *, struct node_st *);
+int8_t number_expr(struct parser_st *, struct node_st *);
 
-int string_expr(struct parser_st *, struct node_st *);
+int8_t string_expr(struct parser_st *, struct node_st *);
 
-int null_expr(struct parser_st *, struct node_st *);
+int8_t null_expr(struct parser_st *, struct node_st *);
 
-int literal_expr(struct parser_st *, struct node_st *);
+int8_t literal_expr(struct parser_st *, struct node_st *);
 
-int atom_expr(struct parser_st *, struct node_st *);
+int8_t atom_expr(struct parser_st *, struct node_st *);
 
-int primary_expr(struct parser_st *, struct node_st *);
+int8_t primary_expr(struct parser_st *, struct node_st *);
 
-int set_primary_expr(struct parser_st *, struct node_st *);
+int8_t set_primary_expr(struct parser_st *, struct node_st *);
 
 #endif //AN_EXPR_H
